Move scripts to a common scripts directory
