+++ backend/core/package.json
@@ -27,7 +27,7 @@
     "typeorm": "ts-node ./node_modules/.bin/typeorm",
     "herokusetup": "node heroku.setup.js",
     "heroku-postbuild": "rimraf dist && nest build && yarn run db:migration:run",
-    "generate:client": "ts-node generate-axios-client.ts",
+    "generate:client": "ts-node scripts/generate-axios-client.ts",
     "migrator": "ts-node migrator.ts"
   },
   "dependencies": {
