+++ backend/core/scripts/generate-axios-client.ts
