+++ backend/core/scripts/listings-importer.ts
@@ -1,8 +1,8 @@
-import * as client from "./client"
+import * as client from "../client"
 import axios from "axios"
-import { serviceOptions } from "./client"
+import { serviceOptions } from "../client"
 import fs from "fs"
-import { ListingStatus } from "./src/listings/types/listing-status-enum"
+import { ListingStatus } from "../src/listings/types/listing-status-enum"
 
 // NOTE: unit's monthlyRent, floor and monthlyIncomeMin have been changed to type string
 // NOTE: in the DB and unit_transformations lib has been adjusted (verify correctness)
