+++ backend/core/src/custom.d.ts
@@ -3,7 +3,7 @@
 declare global {
   declare module "express-serve-static-core" {
     export interface Request {
-      user?: import("./src/user/entities/user.entity").User
+      user?: import("./user/entities/user.entity").User
     }
   }
 }
